Fix cert copying code
the previous code didn't preserve symlinks, and broke when redeployed to the
server. the current code copies everything in /etc/letsencrypt and uses rsync
in both directions so symlinks are preserved. permissions turned out to be a
big pain so just copy them down to kevin:staff and copy them back and chown to
root:root afterwards.